create consultations.js route file
