+++ public/views/dr/index.jade
@@ -19,7 +19,7 @@ block content
 
       br
 
-      div(ng-controller='DrFroms')
+      div(ng-controller='DrForms')
         form.simple-form(novalidate='')
           | Name:
           input(type='text', ng-model='user.name')
